backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic
email-validator
python-multipart
python-dotenv
groq
fpdf
